requirements.txt
torch
timm
opencv-python
numpy
matplotlib
Pillow
tqdm
